3.12/n example overhaul: bollinger -> assets_type_metadata
