+++ docs/content/concepts/types.mdx
@@ -124,4 +124,4 @@ def double_even_with_annotations(num: int) -> int:
 
 ## See it in action
 
-For more examples of the dagster type system, check out our [Bollinger example](https://github.com/dagster-io/dagster/tree/master/examples/bollinger)
+For more examples of the dagster type system, check out our [Type and Metadata example](https://github.com/dagster-io/dagster/tree/master/examples/assets_type_metadata)
