+++ examples/assets_type_metadata/assets_type_metadata/__init__.py
@@ -1,0 +1,2 @@
+from .repository import assets_type_metadata
+from . import lib
