+++ examples/assets_type_metadata/assets_type_metadata/assets/__init__.py
@@ -1,0 +1,1 @@
+
