+++ examples/assets_type_metadata/assets_type_metadata_tests/__init__.py
@@ -1,0 +1,1 @@
+
