+++ examples/assets_type_metadata/scripts/download_stock_data.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+# pylint: disable=print-call
+
+import os
+import sys
+
+import assets_type_metadata.lib as bol
+
+sys.path.append(os.path.join(os.path.dirname(__file__), "../../assets_type_metadata"))
+
+
+path = bol.normalize_path("all_stocks_5yr.csv")
+print(f"Downloading S&P 500 CSV data from {bol.SP500_CSV_URL} to {path}...")
+bol.download_file(bol.SP500_CSV_URL, path)
+print("Successful.")
