+++ examples/assets_type_metadata/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = assets_type_metadata
