+++ examples/assets_type_metadata/tox.ini
@@ -1,0 +1,27 @@
+[tox]
+envlist = py{38, 37},pylint,mypy
+
+[testenv]
+usedevelop = true
+setenv =
+  VIRTUALENV_PIP=21.3.1
+passenv = CI_* COVERALLS_REPO_TOKEN BUILDKITE
+deps =
+  -e ../../python_modules/dagster[mypy,test]
+  -e ../../python_modules/dagit
+  -e ../../python_modules/dagster-graphql
+  -e ../../python_modules/dagster[mypy,test]
+  -e ../../python_modules/libraries/dagster-pandera/
+allowlist_externals =
+  /bin/bash
+commands =
+  !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster -e dagit'
+  pytest -vv
+
+[testenv:mypy]
+commands =
+  mypy --config=../../pyproject.toml --non-interactive --install-types {posargs} .
+
+[testenv:pylint]
+commands =
+  pylint -j0 --rcfile=../../pyproject.toml {posargs} assets_type_metadata assets_type_metadata_tests
