+++ examples/assets_type_metadata/workspace.yaml
@@ -1,0 +1,2 @@
+load_from:
+  - python_package: assets_type_metadata
