--- examples/bollinger/README.md
@@ -1,15 +1,0 @@
-## Bollinger example
-
-A small graph of software-defined assets that compute bollinger bands for S&P 500 prices.
-
-### Loading the example into Dagit
-
-    dagit -w workspace.yaml
-
-### Features demonstrated
-
-- Software-defined assets.
-- Dagster types.
-- `dagster-pandera` integration.
-- Custom `IOManager`.
-- Definition-level metadata on assets.
