--- examples/bollinger/tox.ini
@@ -1,24 +1,0 @@
-[tox]
-envlist = py{39, 38, 37, 36},pylint,mypy
-
-[testenv]
-usedevelop = true
-setenv =
-  VIRTUALENV_PIP=21.3.1
-passenv = CI_* COVERALLS_REPO_TOKEN BUILDKITE SNOWFLAKE_ACCOUNT SNOWFLAKE_USER SNOWFLAKE_PASSWORD
-deps =
-  -e ../../python_modules/dagster[mypy,test]
-  -e ../../python_modules/libraries/dagster-pandera/
-allowlist_externals =
-  /bin/bash
-commands =
-  !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster -e dagit'
-  pytest -vv
-
-[testenv:mypy]
-commands =
-  mypy --config=../../pyproject.toml --non-interactive --install-types {posargs} .
-
-[testenv:pylint]
-commands =
-  pylint -j0 --rcfile=../../pyproject.toml {posargs} bollinger bollinger_tests
