--- examples/bollinger/workspace.yaml
@@ -1,2 +1,0 @@
-load_from:
-  - python_package: bollinger
